fga-eps-mds/2021.1-Oraculo#30 refactor: :recycle: changes component path - adds button type to component
