+++ src/Components/MainButton/index.js
@@ -1,0 +1,9 @@
+import { StyledButton } from "./style";
+
+const MainButton = ({ title, onClick }) => (
+    <StyledButton type="button" onClick={onClick}>
+        {title}
+    </StyledButton>
+);
+
+export default MainButton;
