+++ src/Components/MainButton/style.js
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+
+export const StyledButton = styled.button`
+    height: 2.5rem;
+    width: 19.5rem;
+
+    background-color: #5289b5;
+    border: none;
+    border-radius: 24px;
+    color: #ffffff;
+    cursor: pointer;
+
+    font-family: Montserrat;
+    font-style: normal;
+    font-weight: bold;
+    font-size: 24px;
+    line-height: 20px;
+    text-align: center;
+    justify-content: center;
+`;
